Flyway demonstration
Rather than letting Hibernate create the schema for us, we add flyway
with an initial script that creates the schema and populates it.

This commit also adds the actuator which will add management and devops
features to your application. This namely expose a `/flyway` endpoint you
can query to see the state of the database.