Devtools demonstration
This commit adds devtools, a set of development utilities that improves
your productivity, typically when you're running the application in your
local IDE.

When devtools is enabled, a LiveReload-compatible server is started with
your application so you can connect to it using the LiveReload client.

Development properties are also automatically set. For instance the H2
console is exposed unless specified otherwise so we can see the content
of our in-memory database.