Devtools demonstration (II)
Devtools not only enables nice features in your local environment, it
also includes a LiveReload integration. When your app is started that
way you can connect to it in your browser using the free LiveReload
plugin.

In this commit, we add a new property to our `Person`. This requires a
database change but since we are using flyway, it is very easy to add a
script.

If we save those changes in a running application, devtools will restart
the application for us and refresh the browser when it's done. Because
we use flyway, our schema will be updated transparently.